[LinalgExt] Move some encoding methods to LinalgExt/Utils.
