+++ compiler/src/iree/compiler/Codegen/Common/EncodingUtils.h
@@ -9,7 +9,6 @@
 
 #include "iree/compiler/Dialect/HAL/IR/HALTypes.h"
 #include "iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.h"
-#include "mlir/Dialect/Linalg/IR/Linalg.h"
 #include "mlir/Transforms/DialectConversion.h"
 
 namespace mlir::iree_compiler {
@@ -72,14 +71,6 @@ protected:
 //===---------------------------------------------------------------------===//
 // Utility methods about Encoding.
 //===---------------------------------------------------------------------===//
-
-/// Returns the encoding attribute from the type if there is an encoding.
-/// Otherwise, returns null.
-IREE::LinalgExt::EncodingAttr getEncodingAttr(RankedTensorType type);
-
-/// Returns the ContractionDimensions for the encoding user_indexing_maps.
-FailureOr<linalg::ContractionDimensions>
-getEncodingContractionDims(IREE::LinalgExt::EncodingAttr encoding);
 
 /// Returns the original type that carried by encoding.
 RankedTensorType getOriginalTypeWithEncoding(RankedTensorType type);
